Revert "refactor: :recycle: refatorando criação de usuarios"
