--- .env.exemple
@@ -1,24 +1,0 @@
-APP_API_URL=
-APP_PORT=
-APP_SECRET=
-ENVIRONMENT=
-disk=local
-
-## Postgres DB
-POSTGRES_DB_NAME=
-POSTGRES_DB_HOST=
-POSTGRES_DB_PORT=
-POSTGRES_DB_USERNAME=
-POSTGRES_DB_PASSWORD=
-POSTGRES_DB_DATABASE=
-
-
-## AWS credentials
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_BUCKET=
-AWS_BUCKET_REGION=
-AWS_BUCKET_URL=
-FORGOT_MAIL_URL=
-
-MIGRATION=local
