--- src/adapter/users/rest/controller/createUsers/index.ts
@@ -1,1 +1,0 @@
-export * from './createUserController'
