--- src/adapter/users/rest/presenter/contracts/controller.ts
@@ -1,6 +1,0 @@
-import { IUserDtoInput } from "usecases/system/dtos/user-dto";
-import { HttpResponse } from "./http";
-
-export interface Controller{
-  handle(data:IUserDtoInput): Promise<HttpResponse>
-}
