--- src/adapter/users/rest/presenter/contracts/index.ts
@@ -1,2 +1,0 @@
-export * from './http'
-export * from './controller'
