--- src/adapter/users/rest/router/users.router.ts
@@ -1,1 +1,0 @@
-console.log("users  router")
