+++ src/domain/user/entities/user.spec.ts
