+++ src/domain/user/repository/user-repository.interface.ts
@@ -6,5 +6,4 @@ import { User } from "../entities/user";
 export default interface UserRepositoryInterface
     extends RepositoryInterface<User> {
         Search(entity: string): Promise<void>;
-        findyByEmail(email: string): Promise<User>
     }
