--- src/domain/users/repository/index.ts
@@ -1,1 +1,0 @@
-export * from './user-repository.interface'
