--- src/infra/database/connections.ts
@@ -1,1 +1,0 @@
-console.log("add db connections")
