--- src/infra/database/typeorm/entities/inMemmory/users.ts
@@ -1,12 +1,0 @@
-export class Users{
-  _id?: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: number;
-  bio: string;
-  photo: string;
-  password: string;
-  created_at: Date;
-  updated_at: Date;
-}
