+++ src/main.ts
@@ -1,0 +1,6 @@
+function main() {
+    return 1+2
+}
+
+console.log(main())
+console.log("App running...")
