--- src/main/routes/routes.ts
@@ -1,1 +1,0 @@
-console.log("call all routes here")
