--- src/usecases/users/createUsers/index.ts
@@ -1,1 +1,0 @@
-export * from './user-create.usecase'
